Refactoring and no pointless graph-loading anymore.
