+++ algo/src/main/java/org/neo4j/graphalgo/impl/metaPathComputation/MetaPath.java
@@ -1,4 +1,4 @@
-package org.neo4j.graphalgo;
+package org.neo4j.graphalgo.impl.metaPathComputation;
 
 import org.neo4j.graphalgo.api.*;
 import org.neo4j.graphalgo.core.IdMap;
