+++ tests/src/test/java/org/neo4j/graphalgo/algo/GettingStartedTest.java
@@ -9,7 +9,7 @@ import org.neo4j.graphalgo.TestDatabaseCreator;
 import org.neo4j.graphdb.Transaction;
 import org.neo4j.kernel.api.exceptions.KernelException;
 import org.neo4j.kernel.impl.proc.Procedures;
-import org.neo4j.graphalgo.GettingStartedProc;
+import org.neo4j.graphalgo.metaPathComputationProcs.GettingStartedProc;
 import org.neo4j.kernel.internal.GraphDatabaseAPI;
 
 import java.util.Arrays;
