+++ tests/src/test/java/org/neo4j/graphalgo/impl/metaPathComputationTests/ComputeAllMetaPathsForInstancesTest.java
@@ -1,4 +1,4 @@
-package org.neo4j.graphalgo.impl;
+package org.neo4j.graphalgo.impl.metaPathComputationTests;
 
         import org.junit.AfterClass;
         import org.junit.Before;
@@ -12,7 +12,7 @@ package org.neo4j.graphalgo.impl;
         import org.neo4j.graphdb.*;
         import org.neo4j.kernel.api.exceptions.KernelException;
         import org.neo4j.kernel.impl.proc.Procedures;
-        import org.neo4j.graphalgo.GettingStartedProc;
+        import org.neo4j.graphalgo.metaPathComputationProcs.GettingStartedProc;
         import org.neo4j.kernel.internal.GraphDatabaseAPI;
 
         import java.util.*;
